Fix build for Kafka clients package reorganization
Update imports for LogContext and AppInfoParser, which moved from
org.apache.kafka.common.utils to org.apache.kafka.common.utils.internals.
Add kafka-clients test-fixtures dependency for MockClient, MockTime,
TestUtils, and TestSslUtils, which were split out of the test classifier.
